Implemented CarWash class, Updated Car class.
Added a CarWash class with various different fields, an add_car() method allowing cars to be added to the cars list and a dirty_cars() method returning all cars that need cleaned.

Added size attribute to Car class as the car wash is priced based on size.

Added make_clean() and make_dirty() methods allowing a cars cleanliness status to be updated as needed.